Implemented more complete frame serialisation

+++ headers/common/ze_buffer.h
@@ -50,7 +50,7 @@ struct ZEBuffer
 	f32 PercentageUsed()
 	{
 		zeSize space = Space();
-		return ((capacity - space) / space) * 100.f;
+		return (((f32)capacity - (f32)space) / (f32)space) * 100.f;
 	}
 
     i32 IsValid()
